Petoi Bittle commands

To interact with the robot, the library exposes a PetoiBittle.Command interface and the PetoiBittle.send_command function.

PetoiBittle.Command — Type

Command

An abstract type incapsulated all available commands.

Use

julia> subtypes(PetoiBittle.Command)

to get the list of all available commands.

See also: PetoiBittle.send_command

PetoiBittle.send_command — Function

send_command(connection::Connection, command::Command)

A function to send a command to a Bittle robot through opened PetoiBittle.Connection.

Use

julia> subtypes(PetoiBittle.Command)

to get the list of all available commands.

See also: PetoiBittle.Command

The list of available commands

The full list of available commands for Petoi Robots and their specification can also be found on the Petoi Bittle website in the Serial Protocol section. This library implements a subset of commands. PRs are welcome to add new commands!

To get the list of all available commands implemented in the package use in REPL:

julia> subtypes(PetoiBittle.Command)1-element Vector{Any}:
 PetoiBittle.MoveJoints

Note

This command only works in Julia REPL.

Move joints

PetoiBittle.MoveJoints — Type

MoveJoints(joint_movements_as_named_tuple...)
MoveJoints(single_tuple_of_move_joint_specs)

A PetoiBittle.Command that specifies which joint to move at which angle. Can be constructed from a vararg argument list of named tuples with id and angle keys present or a single tuple containing multiple PetoiBittle.MoveJointSpec.

julia> PetoiBittle.MoveJoints(
           (id = 8, angle = 10),
           (id = 9, angle = 20)
       )
PetoiBittle.MoveJoints{2}((PetoiBittle.MoveJointSpec(8, 10), PetoiBittle.MoveJointSpec(9, 20)))

julia> PetoiBittle.MoveJoints(
           (id = 9, angle = 10),
           (id = 9, angle = 20)
       )
ERROR: Cannot create MoveJoints. Duplicate `id` found `9`. Set `check_unique = false` to skip the check.
[...]

See also: PetoiBittle.MoveJointSpec

PetoiBittle.MoveJointSpec — Type

MoveJointSpec(id::Int, angle::Int)

A command specification for a joint with id id to change its angle to angle. Note, that this structure is not a command by itself. To move a single joint you should still use the PetoiBittle.MoveJoints command.

julia> convert(PetoiBittle.MoveJointSpec, (id = 1, angle = 10))
PetoiBittle.MoveJointSpec(1, 10)

julia> convert(PetoiBittle.MoveJointSpec, (id = 1, ))
ERROR: Cannot convert NamedTuple (id = 1,) to `MoveJointSpec(id = ..., angle = ...)`. Missing key `angle`.
[...]

See also: PetoiBittle.MoveJoints