PetoiBittle

Documentation for PetoiBittle.

PetoiBittle parsers

PetoiBittle.jl implements simple parsers to parse responses from serial ports of PetoiBittle directly from bytes

PetoiBittle.parse_number — Function

parse_number(::Type{T}, bytes)

Parse decimal representation of a number of type T from bytes. Stops at non-digits characters, with the exception of - and ..

julia> PetoiBittle.parse_number(Int, [ 0x34, 0x32 ])
42

julia> PetoiBittle.parse_number(Float64, [ 0x2d, 0x30, 0x2e, 0x35 ])
-0.5

parse_number(bytes)

Short-hand for parse_number(Float64, bytes).

parse_number(::Type{T}, bytes, firstindex, lastindex)

Parse decimal representation of a number of type T from bytes at indices between firstindex and lastindex. Stops at non-digits characters, with the exception of - and ..

Returns (result, nextindex), where result is the actual parsed number and nextindex is either:

an index next to a non-digit character or;

lastindex + 1

julia> PetoiBittle.parse_number(Float64, [ 0x34, 0x32 ], 1, 1)
(4.0, 2)

julia> PetoiBittle.parse_number(Float64, [ 0x34, 0x32 ], 1, 2)
(42.0, 3)

PetoiBittle Constants

PetoiBittle.Constants — Module

A sub-module of PetoiBittle containing some useful constants Constants are united in their respective namespaces.

The available namespaces are:

char: a list of character codes in UInt8

tab: UInt8 code of the tab character ' '

newline: UInt8 code of the newline character '

' - caret: UInt8 code of the caret character ' ' - zero: UInt8 code of the zero character '0' (a digit) - one: UInt8 code of the one character '1' (a digit) - two: UInt8 code of the two character '2' (a digit) - three: UInt8 code of the three character '3' (a digit) - four: UInt8 code of the four character '4' (a digit) - five: UInt8 code of the five character '5' (a digit) - six: UInt8 code of the six character '6' (a digit) - seven: UInt8 code of the seven character '7' (a digit) - eight: UInt8 code of the eight character '8' (a digit) - nine: UInt8 code of the nine character '9' (a digit) - null: UInt8 code of the null character ' ' (a terminator) - dot: UInt8 code of the dot character '.' - minus: UInt8 code of the minus character '-'

julia> PetoiBittle.Constants.char.tab
0x09

PetoiBittle.Constants

PetoiBittle.parse_number